split figure into static and interactive, better separation of toolbar
