+++ src/plopp/__init__.py
@@ -10,7 +10,7 @@ plt.ioff()
 from .graph import show_graph
 from .plot import Plot
 from .model import Node, node, input_node
-from .figure import Figure
+from .figure import figure
 from . import widgets
 from .wrappers import plot
