+++ src/plopp/wrappers.py
@@ -2,7 +2,7 @@
 # Copyright (c) 2022 Scipp contributors (https://github.com/scipp)
 
 from .model import input_node
-from .figure import Figure
+from .figure import figure
 from .tools import number_to_variable
 
 from scipp import Variable, DataArray, Dataset, arange, to_unit
@@ -61,7 +61,7 @@ def plot(obj: Union[VariableLike, Dict[str, VariableLike]],
          title: str = None,
          vmin: Variable = None,
          vmax: Variable = None,
-         **kwargs) -> Figure:
+         **kwargs):
     """Plot a Scipp object.
 
     Parameters
@@ -124,6 +124,6 @@ def plot(obj: Union[VariableLike, Dict[str, VariableLike]],
     }
     if isinstance(obj, (dict, Dataset)):
         nodes = [input_node(_preprocess(item, crop=crop)) for item in obj.values()]
-        return Figure(*nodes, **all_args)
+        return figure(*nodes, **all_args)
     else:
-        return Figure(input_node(_preprocess(obj, crop=crop)), **all_args)
+        return figure(input_node(_preprocess(obj, crop=crop)), **all_args)
